update module name
