+++ lib/templates/gametemplate.js
@@ -102,9 +102,7 @@ class GameTemplate extends ModTemplate {
 		};
 
 		this.recordOptions = {
-			container: "#main",
-			useMicrophone: true,
-			streams: null,
+			container: "body",
 			callbackAfterRecord: null
 		}
 
@@ -206,6 +204,7 @@ class GameTemplate extends ModTemplate {
 		this.card_back = 'red_back.png';
 
 		this.publisher_message = '';
+
 
 		this.time = {};
 		this.time.last_received = 0; //For when we last received control of game
