+++ mods/screenrecord/web/css/record.css
